bugfixes and some cleanup to be able to run the script

+++ nebterpolator/smoothing.py
@@ -107,7 +107,7 @@ def window_smooth(signal, window_len=11, window='hanning'):
         w = getattr(np, window)(window_len)
     output = np.convolve(w / w.sum(), padded, mode='valid')
 
-    return output[(window_len / 2):-(window_len / 2)]
+    return output[int(window_len / 2):-int(window_len / 2)]
 
 
 def buttersworth_smooth(signal, width=11, order=3):
